refactor(common): move routing evidence shapes to memex_common.lint
Per review: the skill (and eval) should not depend on memex_core just for the
routing-proposal evidence shapes, and these are wire-contract models that belong
beside LintProposal/ProposedAction. Move CandidateEvidence/RouteEvidence/
NoFitEvidence from memex_core.services.proposal_actions._routing_evidence into
memex_common.lint (the SSOT for lint-proposal wire shapes, which must not import
core) and delete the core module. Repoint the eval scenario and the route-action
test at memex_common.lint. The triage-inbox skill now imports them from common,
dropping its memex_core dependency.
